+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/recipe/VanillaRecipeCache.java
@@ -4,6 +4,7 @@ import com.codetaylor.mc.artisanworktables.api.internal.recipe.*;
 import com.codetaylor.mc.artisanworktables.api.internal.reference.EnumTier;
 import com.codetaylor.mc.artisanworktables.api.recipe.ArtisanRecipe;
 import com.codetaylor.mc.artisanworktables.api.recipe.IArtisanRecipe;
+import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktablesConfig;
 import com.codetaylor.mc.artisanworktables.modules.worktables.integration.crafttweaker.CTArtisanIngredient;
 import com.codetaylor.mc.artisanworktables.modules.worktables.integration.crafttweaker.CTArtisanRecipe;
 import com.codetaylor.mc.artisanworktables.modules.worktables.tile.spi.TileEntityBase;
@@ -54,21 +55,20 @@ public class VanillaRecipeCache {
   @Nullable
   public synchronized static IArtisanRecipe getArtisanRecipe(InventoryWrapper inventoryWrapper, World world) {
 
+    if (inventoryWrapper.isEmpty()) {
+      // If the crafting grid is empty, we don't even try matching a recipe.
+      return null;
+    }
+
     IRecipe recipe;
 
     recipe = LAST_RECIPE.get()[0];
 
     if (recipe == null || !recipe.matches(inventoryWrapper, world)) {
-
-      if (!inventoryWrapper.isEmpty()) {
-        recipe = CraftingManager.findMatchingRecipe(inventoryWrapper, world);
-
-      } else {
-        recipe = null;
-      }
+      recipe = CraftingManager.findMatchingRecipe(inventoryWrapper, world);
     }
 
-    if (recipe == null) {
+    if (recipe == null || ModuleWorktablesConfig.isRecipeBlacklisted(recipe.getRegistryName())) {
       return null;
     }