+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/tile/spi/TileEntityBase.java
@@ -499,6 +499,11 @@ public abstract class TileEntityBase
   @Nullable
   public IArtisanRecipe getRecipe(@Nonnull EntityPlayer player) {
 
+    if (this.craftingMatrixHandler.isEmpty()) {
+      // If the crafting grid is empty, we don't even try matching a recipe.
+      return null;
+    }
+
     FluidStack fluidStack = this.getTank().getFluid();
 
     if (fluidStack != null) {